Imported Upstream version 3.6.1

+++ tests/load/lastseen_threaded_load.c
@@ -3,11 +3,12 @@
 #include <dbm_api.h>
 #include <process_lib.h>                               /* GracefulTerminate */
 #include <mutex.h>                                     /* ThreadLock */
+#include <misc_lib.h>                                  /* xclock_gettime */
 
 #include <libgen.h>                                             /* basename */
 
 
-#define NSEC 10                            /* how long to run each loop */
+unsigned int ROUND_DURATION = 10;          /* how long to run each loop */
 #define NHOSTS 5000                        /* how many hosts to store in db */
 #define MAX_NUM_THREADS 10000
 #define MAX_NUM_FORKS   10000
@@ -110,7 +111,7 @@ static void thread_exit_clean()
     /* Signal that we finished. */
     ThreadLock(&end_mtx);
     FINISHED_THREADS++;
-    if (FINISHED_THREADS > TOTAL_NUM_THREADS)
+    if (FINISHED_THREADS >= TOTAL_NUM_THREADS)
     {
         pthread_cond_signal(&end_cond);
     }
@@ -119,15 +120,15 @@ static void thread_exit_clean()
 
 void *lastsaw_worker_thread(void *arg)
 {
-    intptr_t thread_id = (intptr_t) arg;
+    int thread_id = (intptr_t) arg;
 
     size_t i = 0;
     while (!DONE)
     {
         char hostkey[50];
-        snprintf(hostkey, sizeof(hostkey), "SHA-%040zx", i);
+        xsnprintf(hostkey, sizeof(hostkey), "SHA-%040zx", i);
         char ip[50];
-        snprintf(ip, sizeof(ip), "250.%03zu.%03zu.%03zu",
+        xsnprintf(ip, sizeof(ip), "250.%03zu.%03zu.%03zu",
                  i / (256*256), (i / 256) % 256, i % 256);
 
         UpdateLastSawHost(hostkey, ip,
@@ -146,7 +147,7 @@ void *lastsaw_worker_thread(void *arg)
 
 void *keycount_worker_thread(void *arg)
 {
-    intptr_t id = (intptr_t) arg;
+    int id = (intptr_t) arg;
 
     while (!DONE)
     {
@@ -160,7 +161,7 @@ void *keycount_worker_thread(void *arg)
 
 void *scanlastseen_worker_thread(void *arg)
 {
-    intptr_t id = (intptr_t) arg;
+    int id = (intptr_t) arg;
 
     while (!DONE)
     {
@@ -179,12 +180,14 @@ void *scanlastseen_worker_thread(void *arg)
 
 unsigned long child_COUNTER;
 int PIPE_FD[2];
+FILE *PARENT_INPUT;
 bool child_START = false;
 
-void print_progress_handler(int signum ARG_UNUSED)
+void print_progress_sighandler(int signum ARG_UNUSED)
 {
-    printf("%5lu", child_COUNTER); putc('\0', stdout);
-    fflush(stdout);
+    fprintf(PARENT_INPUT, "%5lu", child_COUNTER);
+    putc('\0', PARENT_INPUT);
+    fflush(PARENT_INPUT);
 
     child_COUNTER = 0;
 }
@@ -202,11 +205,6 @@ void worker_process()
     struct sigaction new_handler;
     int ret;
 
-    /* 0. Output stdout to pipe. */
-    close(1);
-    dup(PIPE_FD[1]);
-    close(PIPE_FD[1]);
-
     /* 1a. Register SIGUSR1 handler so that we start/finish the test */
     new_handler = (struct sigaction) { .sa_handler = startstop_handler };
     sigemptyset(&new_handler.sa_mask);
@@ -218,7 +216,7 @@ void worker_process()
     }
 
     /* 1b. Register SIGUSR2 handler so that we report progress when pinged */
-    new_handler = (struct sigaction) { .sa_handler = print_progress_handler };
+    new_handler = (struct sigaction) { .sa_handler = print_progress_sighandler };
     sigemptyset(&new_handler.sa_mask);
     ret = sigaction(SIGUSR2, &new_handler, NULL);
     if (ret != 0)
@@ -254,9 +252,9 @@ void worker_process()
         else
         {
             char hostkey[50];
-            snprintf(hostkey, sizeof(hostkey), "SHA-%040zx", child_COUNTER);
+            xsnprintf(hostkey, sizeof(hostkey), "SHA-%040lx", child_COUNTER);
             char ip[50];
-            snprintf(ip, sizeof(ip), "250.%03zu.%03zu.%03zu",
+            xsnprintf(ip, sizeof(ip), "250.%03lu.%03lu.%03lu",
                       child_COUNTER / (256*256),
                      (child_COUNTER / 256) % 256,
                       child_COUNTER % 256);
@@ -276,23 +274,23 @@ void worker_process()
 
 
 void spawn_worker_threads(void *(*worker_routine) (void *),
-                           int num_threads, const char *description)
+                          int num_threads, const char *description)
 {
     pthread_t tid[num_threads];
 
     printf("Spawning %s worker threads: ", description);
-    for (intptr_t i = 0; i < num_threads; i++)
+    for (int i = 0; i < num_threads; i++)
     {
         int ret = pthread_create(&tid[i], NULL,
-                                 worker_routine, (void *) i);
+                                 worker_routine, (void *)(intptr_t) i);
         if (ret != 0)
         {
-            fprintf(stderr, "pthread_create(%jd): %s",
-                    (intmax_t) i, GetErrorStrFromCode(ret));
+            fprintf(stderr, "pthread_create(%d): %s",
+                    i, GetErrorStrFromCode(ret));
             exit(EXIT_FAILURE);
         }
 
-        printf("%lu ", i+1);
+        printf("%i ", i+1);
     }
     printf("done!\n");
 }
@@ -342,58 +340,123 @@ void print_progress(int lastsaw_num_threads, int keycount_num_threads,
     }
 }
 
+void print_usage(const char *argv0)
+{
+        printf("\
+\n\
+Usage:\n\
+	%s [options] LASTSAW_NUM_THREADS [KEYCOUNT_NUM_THREADS [SCAN_NUM_THREADS]]\n\
+\n\
+This program creates many threads and optionally many processes stressing the\n\
+lastseen database.\n\
+\n\
+Options:\n\
+	-d N:	Duration of each round of testing in seconds (default is 10s)\n\
+	-c N:	After finishing all rounds with threads, N spawned child\n\
+		processes shall apply a mixed workload to the database each one\n\
+		for another round (default is 0, i.e. don't fork children)\n\
+\n",
+               argv0);
+}
+
 void parse_args(int argc, char *argv[],
                 int *lastsaw_num_threads, int *keycount_num_threads,
-                int *scanlastseen_num_threads)
+                int *scanlastseen_num_threads, int *num_forked_children)
 {
     *lastsaw_num_threads      = 0;
     *keycount_num_threads     = 0;
     *scanlastseen_num_threads = 0;
+    *num_forked_children = 0;
 
-    if (argc >= 2)
+    int i = 1;
+    while (i < argc && argv[i][0] == '-')
     {
-        sscanf(argv[1], "%d", lastsaw_num_threads);
-    }
-    if (argc >= 3)
-    {
-        sscanf(argv[2], "%d", keycount_num_threads);
-    }
-    if (argc >= 4)
-    {
-        sscanf(argv[3], "%d", scanlastseen_num_threads);
+        switch (argv[i][1])
+        {
+        case 'd':
+        {
+            i++;
+            int N = 0;
+            int ret = sscanf((argv[i] != NULL) ? argv[i] : "",
+                             "%d", &N);
+            if (ret != 1 || N <= 0)
+            {
+                print_usage(basename(argv[0]));
+                exit(EXIT_FAILURE);
+            }
+
+            ROUND_DURATION = N;
+            break;
+        }
+        case 'c':
+        {
+            i++;
+            int N = -1;
+            int ret = sscanf((argv[i] != NULL) ? argv[i] : "",
+                             "%d", &N);
+            if (ret != 1 || N < 0)
+            {
+                print_usage(basename(argv[0]));
+                exit(EXIT_FAILURE);
+            }
+
+            *num_forked_children = N;
+            break;
+        }
+        default:
+            print_usage(basename(argv[0]));
+            exit(EXIT_FAILURE);
+        }
+
+        i++;
     }
 
+    /* Last 3 arguments */
+
+    if (i < argc)
+    {
+        sscanf(argv[i], "%d", lastsaw_num_threads);
+    }
+    if (i + 1 < argc)
+    {
+        sscanf(argv[i + 1], "%d", keycount_num_threads);
+    }
+    if (i + 2 < argc)
+    {
+        sscanf(argv[i + 2], "%d", scanlastseen_num_threads);
+    }
+
+    /* lastsaw_num_threads is the only /mandatory/ argument. */
     if (*lastsaw_num_threads  <= 0 || *lastsaw_num_threads  > MAX_NUM_THREADS)
     {
-        printf("\
-Usage:\n\
-	%s LASTSAW_NUM_THREADS [KEYCOUNT_NUM_THREADS [SCAN_NUM_THREADS]]\n\
-\n\
-This program creates a parent process stressing lastseen. After several seconds\n\
-creating the threads you defined, it also spawns a separate process doing things\n\
-to the same database.\n\n",
-               basename(argv[0]));
-
+        print_usage(basename(argv[0]));
         exit(EXIT_FAILURE);
+    }
+
+    if (*num_forked_children > 1)                              /* TODO FIX! */
+    {
+        printf("WARNING: Currently only one forked child is supported TODO FIX!\n");
+        *num_forked_children = 1;
     }
 }
 
 
 int main(int argc, char *argv[])
 {
+    int ret;
     int lastsaw_num_threads, keycount_num_threads, scanlastseen_num_threads;
-    const int total_num_children = 1;
+    int num_forked_children;
 
     LogSetGlobalLevel(LOG_LEVEL_DEBUG);
 
     parse_args(argc, argv,
                &lastsaw_num_threads, &keycount_num_threads,
-               &scanlastseen_num_threads);
+               &scanlastseen_num_threads, &num_forked_children);
     TOTAL_NUM_THREADS =
         lastsaw_num_threads + keycount_num_threads + scanlastseen_num_threads;
 
 
-    snprintf(CFWORKDIR, sizeof(CFWORKDIR),
+    xsnprintf(CFWORKDIR, sizeof(CFWORKDIR),
              "/tmp/lastseen_threaded_load.XXXXXX");
     char *retp = mkdtemp(CFWORKDIR);
     if (retp == NULL)
@@ -402,43 +465,60 @@ int main(int argc, char *argv[])
         exit(EXIT_FAILURE);
     }
     printf("Work directory: %s\n", CFWORKDIR);
-    /* char envvar[CF_BUFSIZE]; */
-    /* snprintf(envvar, sizeof(envvar), "TEST_CFENGINE_OVERRIDE_WORKDIR=%s", */
-    /*          CFWORKDIR); */
-    /* putenv(envvar); */
 
     START_TIME = time(NULL);
 
 
     /* === SPAWN A CHILD PROCESS FOR LATER === */
 
-    pipe(PIPE_FD);
-    pid_t child = fork();
-    if (child == -1)
+    pid_t child;
+    if (num_forked_children > 0)
     {
-        perror("fork");
-        exit(EXIT_FAILURE);
-    }
-    else if (child == 0)
-    {
-        close(PIPE_FD[0]);
-        worker_process();
-        exit(EXIT_SUCCESS);
+        ret = pipe(PIPE_FD);
+        if (ret != 0)
+        {
+            perror("pipe");
+            exit(EXIT_FAILURE);
+        }
+
+        child = fork();
+        if (child == -1)
+        {
+            perror("fork");
+            exit(EXIT_FAILURE);
+        }
+        else if (child == 0)                                    /* child */
+        {
+            /* We only write to the pipe. */
+            close(PIPE_FD[0]);
+            /* Connect pipe to a FILE to talk to parent via fprintf(). */
+            PARENT_INPUT = fdopen(PIPE_FD[1], "w");
+            if (PARENT_INPUT == NULL)
+            {
+                perror("child fdopen");
+                exit(EXIT_FAILURE);
+            }
+
+            worker_process();
+            exit(EXIT_SUCCESS);
+        }
+
+        /* We only read from the pipe. */
+        close(PIPE_FD[1]);
+        CHILDREN_OUTPUTS[0] = PIPE_FD[0];
     }
 
-    /* For now we support only one child process. */
-    CHILDREN_OUTPUTS[0] = PIPE_FD[0];
-    close(PIPE_FD[1]);
-    printf("Operations per second:\n\n");
 
-    /* === CREATE WORKER THREADS === */
+    printf("Showing number of operations per second:\n\n");
+
+    /* === CREATE lastsaw() WORKER THREADS === */
 
     spawn_worker_threads(lastsaw_worker_thread, lastsaw_num_threads,
                          "UpdateLastSawHost()");
 
-    /* === PRINT PROGRESS FOR NSEC SECONDS === */
+    /* === PRINT PROGRESS FOR ROUND_DURATION SECONDS === */
 
-    for (int i = 0; i < NSEC; i++)
+    for (int i = 0; i < ROUND_DURATION; i++)
     {
         sleep(1);
         print_progress(lastsaw_num_threads, 0, 0, 0);
@@ -452,8 +532,8 @@ int main(int argc, char *argv[])
         spawn_worker_threads(keycount_worker_thread, keycount_num_threads,
                              "LastSeenHostKeyCount()");
 
-        /* === PRINT PROGRESS FOR NSEC SECONDS === */
-        for (int i = 0; i < NSEC; i++)
+        /* === PRINT PROGRESS FOR ROUND_DURATION SECONDS === */
+        for (int i = 0; i < ROUND_DURATION; i++)
         {
             sleep(1);
             print_progress(lastsaw_num_threads, keycount_num_threads, 0, 0);
@@ -468,8 +548,8 @@ int main(int argc, char *argv[])
         spawn_worker_threads(scanlastseen_worker_thread, scanlastseen_num_threads,
                              "ScanLastSeenQuality()");
 
-        /* === PRINT PROGRESS FOR NSEC SECONDS === */
-        for (int i = 0; i < NSEC; i++)
+        /* === PRINT PROGRESS FOR ROUND_DURATION SECONDS === */
+        for (int i = 0; i < ROUND_DURATION; i++)
         {
             sleep(1);
             print_progress(lastsaw_num_threads, keycount_num_threads,
@@ -480,24 +560,28 @@ int main(int argc, char *argv[])
 
     /* === START CHILD PROCESS WORK === */
 
-    printf("Forking child doing mix of operations\n");
-    kill(child, SIGUSR1);
-
-    /* === PRINT PROGRESS FOR NSEC SECONDS === */
-
-    for (int i = 0; i < NSEC; i++)
+    if (num_forked_children > 0)
     {
-        sleep(1);
-        kill(child, SIGUSR2);                /* receive progress from child */
-        print_progress(lastsaw_num_threads, keycount_num_threads,
-                       scanlastseen_num_threads, 1);
-        putc('\n', stdout);
+        printf("Doing mix of operations in forked children\n");
+        kill(child, SIGUSR1);
+
+        /* === PRINT PROGRESS FOR ROUND_DURATION SECONDS === */
+
+        for (int i = 0; i < ROUND_DURATION; i++)
+        {
+            sleep(1);
+            kill(child, SIGUSR2);                /* receive progress from child */
+            print_progress(lastsaw_num_threads, keycount_num_threads,
+                           scanlastseen_num_threads, 1);
+            putc('\n', stdout);
+        }
+
+        kill(child, SIGUSR1);                     /* signal child to finish */
     }
 
-    /* === TEST FINISHED === */
+    /* === TEST FINISHED, signal threads to exit === */
 
     DONE = true;
-    kill(child, SIGUSR1);
 
     /* === WAIT AT MOST 30 SECONDS FOR EVERYBODY TO FINISH === */
 
@@ -508,10 +592,14 @@ int main(int argc, char *argv[])
     time_t seconds_waited = 0;
     ThreadLock(&end_mtx);
     while (!(    FINISHED_THREADS == TOTAL_NUM_THREADS
-             && finished_children == total_num_children)
+             && finished_children == num_forked_children)
            && seconds_waited < 30)
     {
-        struct timespec ts = { .tv_sec = 1 };
+        struct timespec ts;
+        xclock_gettime(CLOCK_REALTIME, &ts);
+
+        /* Wait at most 1s for the thread to signal us before looping over. */
+        ts.tv_sec++;
         if (FINISHED_THREADS < TOTAL_NUM_THREADS)
         {
             pthread_cond_timedwait(&end_cond, &end_mtx, &ts);
@@ -534,7 +622,7 @@ int main(int argc, char *argv[])
     /* === CLEAN UP TODO register these with atexit() === */
 
     int retval = EXIT_SUCCESS;
-    if (finished_children != total_num_children)
+    if (finished_children != num_forked_children)
     {
         fprintf(stderr,
                 "Forked child seems to be still alive, killing it!\n");
@@ -555,9 +643,10 @@ int main(int argc, char *argv[])
         printf("DONE!\n\n");
     }
 
-    char cmd[CF_BUFSIZE];
-    snprintf(cmd, CF_BUFSIZE, "rm -rf '%s'", CFWORKDIR);
+    char *cmd;
+    xasprintf(&cmd, "rm -rf '%s'", CFWORKDIR);
     system(cmd);
+    free(cmd);
 
     return retval;
 }
